move RangeUtil to yggdrasil
